requirements.txt
streamlit
opencv-python-headless
insightface==0.7.3
onnxruntime
requests
Pillow
